It appears that the bug in dumpi_to_graph causing it to deadlock when processing traces from miniAMR is related to miniAMR's use of user-defined communicators in its load-balancing routines. (Specifically, some of the p2p communication in rcb.c)
With this in mind, dumpi_to_graph needs to do remote message matching
exchange in a way that respects user-defined communicators. To do this,
we need to track user-defined communicators as the traces are parsed.

This commit adds stub callbacks to do this, at least to the degree
necessary for miniAMR. We add callbacks for:
- MPI_Comm_split
- MPI_Comm_rank
- MPI_Comm_size